requirements.txt
Flask==0.11.1
google-cloud
gunicorn==19.6.0
oauth2client==3.0.0
Flask-SQLAlchemy==2.1
PyMySQL==0.7.9
Flask-PyMongo==0.4.1
PyMongo==3.3.0
six==1.10.0
google-api-python-client
google-cloud-translate
wave
# urllib.request
